refactor: Move domain layer classes into designated packages
