--- app/src/main/java/me/worric/kotlinplayground/domain/Command.kt
@@ -1,5 +1,0 @@
-package me.worric.kotlinplayground.domain
-
-interface Command<out T> {
-    fun execute(): T
-}
