+++ app/src/main/java/me/worric/kotlinplayground/domain/commands/Command.kt
@@ -1,0 +1,5 @@
+package me.worric.kotlinplayground.domain.commands
+
+interface Command<out T> {
+    fun execute(): T
+}
